move tests; update docs
